refactor(datasets): move export and FR translation into provenance modal
Dataset card header now has only Provenance + Delete (private). The
provenance modal footer hosts the Export dropdown and Add/Edit French
Translation button (private only), keeping all dataset actions in one place.

- DatasetProvenanceModal: add export dropdown, FR translation button,
  and nested DatasetTranslationModal; pull in fetchTranslation +
  authHeader + metadataUrl + downloadJson from Dataset.js
- Dataset.js: remove export dropdown, FR translation button,
  DatasetTranslationModal, and all related state/handlers/imports
